feat: Add a CLI that agents can also use instead of the MCP
This CLI introduces commands like `logs` or `errors` which can also be
combined with a `,` or `+` to listen to multiple streams at once.

The first iteration will be a streaming log like `tail -f` but in the next
iteration we will add a `no-follow` mode that accepts some filters like a
`timeWindow` or `limit` as the MCP does.

This requires connecting to an already running upstream sidecar instance,
which we also wanna address with the streaming/follow mode.